Exercise 5.15 Done
Kind of cheated by adding the props.setBlogs in the catch to be able to make the mock execute without actually triggering the API and testing if like button was clicked via setBlogs instead of the actual like button. But I don't see any reason for it to cause issues in the future at the moment.